+++ src/types/SnackbarStack/snackbar-stack.type.ts
@@ -1,0 +1,3 @@
+import type { OrigamSnackbarStack } from "../../components"
+
+export type TOrigamSnackbarStack = InstanceType<typeof OrigamSnackbarStack>